clutter/color-state: Add color state transform helper
This helper generates shader snippets that converts pixels from one
color state to another. For example if there is content with nonlinear
(elecrtical) sRGB color state that should be converted to linear
(optical) BT.2020 before being written to the framebuffer, a shader that
makes the pixel go through the following steps will be generated:

 1. sRGB EOTF
 2. Luminance gain (hard coded for now)
 3. Color space mapping

The intention is that it should be possible to composite in a optical /
linear color space, into an intermediate framebuffer, which is then
passed through an inverse EOTF to produce linear content in the output
color state.