keybindings: Guard X11 calls
This is absolutely not a great commit. As the keybindings could use a decent
refactoring, see https://gitlab.gnome.org/GNOME/mutter/-/issues/3363 for more details.
And as this is the last remaining item in getting a wayland-only build, I think
it is a good 'comporomise' for the time being and makes it clear which bits of code are
used in x11-only code paths.