clutter/color-state: Introduce ClutterColorStateIcc
This new color-state parses an ICC profile using lcms2.

When parsing it, generates eotf/inv_eotf profiles that will be used if
the transform is required to go from linear / to linear. This allows
blending in linear.

The transform snippet is always the same: a 3D LUT with tetrahedral
interpolation.

The 3D LUT data is generated using a lcms Transfom from origin ICC Profile to
target ICC Profile.

The 3D LUT is uploaded as a 2D texture (lut_size x (lut_size x lut_size)).

With Cogl, textures can only be updated as layers. Working with layers
is a bit more rigid than working with gl Textures. They already have
predefined snippets and are usually chained with a previous layer until
setting the last layer to cogl_color_out.
To workaround this:
  - This texture is uploaded as the last layer (layer
    index 10) to not conflict with other layers.
  - Cogl implicitly sets the texture name considering layer index, so
    it has to be accessed as "cogl_sampler10".
  - The predefined snippet is: give the result from the previous layer.