clutter/color-state: Add do_transform method
This is a CPU-based transformation method that performs the
same transformation implemented on GPU shaders.

The transformation gets an array of pixels in RGB float format and
returns an array of pixels transformed.

The returned pixels can have different formats:
  - RGBA_FP_32323232:
    Used for testing.
  - RGBX_FP_16161616:
    If half float is supported, used to upload 3D LUT samples.
  - RGBX_8888:
    If half float is unsupported, used to upload 3D LUT samples.

This will be used in the next commit to compare the results
between CPU and GPU, validating the shader implementations.