cursor-renderer: Split update cursor_renderer position and cursor
cursor_render is only updated on Wayland backend, because on
X11 backend some issues appear.

This made the cursor on X11 screen recording appear on the same position,
although its shapes are updated.

To fix this, update only the cursor_render position on X11 backend and
on wayland backend update the cursor too.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/3182