clutter/clutter-color-state: Use 64 bits on id
The color-state id was generated from a counter of 32 bits.
There was an improbable case where the ids counter could be wrapped
around.

Using a counter of 64 bits makes the wrapping around harder to
happen.
This fits with the next version of the color management protocol where
image_description ids will be of 64 bits.

The current version of color management protocol allows recycling used ids,
implement that to avoid the wrap around issue.

Now each color-state has a unique recyclabe image description id for the
color management protocol, which is not the same as its internal 64 bit id.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/3947