cogl/pipeline: When comparing layers, check also the layer index
When checking if some parts of a pipeline can be reused, the layer_index
wasn't being checked.

On Cogl journal flush, the attributes of the pipeline are created
considering the layer index, i.e. cogl_tex_coord5_in. These attributes
end up being searched with glGetAttribLocation and enabled if found or disabled
if not.

When a layer from a pipeline is considered to be reused in a different one,
if the layer indices differ, the attributes won't be found and
would end up being disabled, so that pipeline won't work.