clutter/color-state: Add snippet to_XYZ, and from_XYZ
This will be used when two color-states are of different types, i.e. one
parametric type and the other ICC type.

In that case, the transformation is done using an intermediate space between
each color state. This intermediate space is XYZ, linear and uses
D50 as white point, it also has the default SDR luminances.

The transformation is split into to:
  - One transformation from source color state to this intermediate space.
  - Another transformation from this intermediate space to the target
    color state.